Fix warnings under strictest Eclipse compiler settings

+++ src/org/lwjgl/demo/opengl/raytracing/AtomicDemo.java
@@ -483,7 +483,7 @@ public class AtomicDemo {
 		 * = 0.0 - use only the new frame
 		 * > 0.0 - blend between old frame and new frame
 		 */
-		float blendFactor = (float) frameNumber / ((float) frameNumber + 1.0f);
+		float blendFactor = frameNumber / (frameNumber + 1.0f);
 		glUniform1f(blendFactorUniform, blendFactor);
 		glUniform1i(bounceCountUniform, bounceCount);
