+++ src/org/lwjgl/demo/opengl/raytracing/Demo33Ubo.java
@@ -440,7 +440,7 @@ public class Demo33Ubo {
 		 * compute the blend factor between old frame and new frame. 0.0 - use
 		 * only the new frame > 0.0 - blend between old frame and new frame
 		 */
-		float blendFactor = (float) frameNumber / ((float) frameNumber + 1.0f);
+		float blendFactor = frameNumber / (frameNumber + 1.0f);
 		glUniform1f(blendFactorUniform, blendFactor);
 		glUniform1i(bounceCountUniform, bounceCount);
